Update Aquaponics.md
The overarching theme of my updates is that our charter persists beyond the current year, membership, and initiatives. For example, our library association is not critical to our persistent purpose. If the library were to stop sponsoring Aquaponics, we would not disappear. So the charter should be able to withstand changes in month-to-month and ideally year-to-year situations. I made an exception for the material and space section - we'll need to update that if it becomes inaccurate.

Other than that, most edits are done with an eye towards concision.

One specific edit to content: with leadership changes, it makes sense to pick one protocol, rather than two options. I edited it to be co-leaders selecting new members.